gateway/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
starlette==0.47.2

clickhouse_driver
confluent_kafka
clickhouse-connect==0.8.18
psycopg2-binary==2.9.10
redis==6.4.0


aiokafka==0.12.0


aiohttp==3.12.15
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
h11==0.16.0
httpx==0.27.2
idna==3.10
multidict==6.6.3
requests==2.32.4
sniffio==1.3.1
urllib3==2.5.0
yarl==1.20.1


anyio==4.10.0
asyncio==4.0.0


aiogram==3.21.0
magic-filter==1.0.12
aiofiles==24.1.0


annotated-types==0.7.0
click==8.1.8
exceptiongroup==1.3.0
loguru==0.7.3
lz4==4.4.4
packaging==25.0
platformdirs==4.3.8
propcache==0.3.2
psutil==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
shortuuid==1.0.13
typing-extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
validators==0.35.0
zipp==3.23.0
zstandard==0.24.0


prometheus_client==0.23.1


PyJWT==2.10.1

pydantic_settings